+++ snake_web/constants/DSnakeWeb.py
@@ -3,4 +3,3 @@ from typing import Final
 
 class DSnakeWeb:
     VERSION: Final[str] = "1.4.0"
-    POLL_INTERVAL: Final[int] = 300