+++ snake_web/constants/DSnakeWeb.py
@@ -2,4 +2,4 @@ from typing import Final
 
 
 class DSnakeWeb:
-    VERSION: Final[str] = "1.4.0"
+    VERSION: Final[str] = "1.5.0"